Safari-611.1.10/LayoutTests/media/video-currentsrc-cleared-expected.txt


Check that 'currentsrc' is cleared when there is no media resource.

EVENT(canplaythrough)
EXPECTED (video.currentSrc.indexOf("content/test") > '-1') OK
RUN(video.src = "")

EVENT(error)
EXPECTED (video.currentSrc == '') OK

END OF TEST
